+++ src/cpp_audio/synths/MonauralBeatStereoAmps.cpp
@@ -1,5 +1,5 @@
 #include "MonauralBeatStereoAmps.h"
-#include "AudioUtils.h"
+#include "../utils/AudioUtils.h"
 #include <cmath>
 #include <algorithm>
