+++ src/cpp_audio/synths/StereoAMIndependent.cpp
@@ -1,5 +1,5 @@
 #include "StereoAMIndependent.h"
-#include "AudioUtils.h"
+#include "../utils/AudioUtils.h"
 #include <cmath>
 
 using namespace juce;
